Add input fields for arguments to deebot-command nodes, moved commands.js to be accessible in editor
